% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blb_linear_regression.R
\name{blb_linear_regression}
\alias{blb_linear_regression}
\title{Function for BLB in linear regression}
\usage{
blb_linear_regression(X, Y, b, s, r)
}
\arguments{
\item{X}{numeric matrix of covariates}

\item{Y}{numeric vector of responses}

\item{b}{numeric input for the subset size}

\item{s}{numeric input for the number of sampled subsets}

\item{r}{numeric input for the number of Monte Carlo iterations}
}
\value{
blb estimate of coefficients and sd of the estimates
}
\description{
Implements the BLB algorithm in linear regression
}
\keyword{linear}
\keyword{regression}
