% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blb_quantile_regression.R
\name{blb_quantile_regression}
\alias{blb_quantile_regression}
\title{Function for BLB in quantile regression}
\usage{
blb_quantile_regression(X, Y, b, s, r, q = 0.5)
}
\arguments{
\item{X}{numeric matrix of covariates}

\item{Y}{numeric vector of responses}

\item{b}{numeric input for the subset size}

\item{s}{numeric input for the number of sampled subsets}

\item{r}{numeric input for the number of Monte Carlo iterations}

\item{q}{numeric input for the quantile for quantile regression}
}
\value{
blb estimate of coefficients and sd of the estimates
}
\description{
Implements the BLB algorithm in quantile regression
}
\keyword{quantile}
\keyword{regression}
