% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_regr.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(x, y, B, n)
}
\arguments{
\item{B}{the number of bootstrap replications}

\item{n}{the number of resamples to generate}

\item{data}{the original sample, as a vector}

\item{data}{is a vector}
}
\value{
vector with the standard deviation of the regression coeffiecients
}
\description{
Bootstrap
}
\examples{
bootstrap(,,,,,)
}

