% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_multidim.R
\name{bootstrap_multidim}
\alias{bootstrap_multidim}
\title{Bootstraps for a regression coefficients}
\usage{
bootstrap_multidim(X, y, r, n)
}
\arguments{
\item{X}{The data as a matrix}

\item{y}{The response variable for the regression}

\item{r}{the number of bootstrap resamples}

\item{n}{the number of data points in each resample (which is also the number of data points in the full data set)}
}
\value{
vector with the standard deviation of the regression coeffiecients
}
\description{
Bootstrap functions that performs regression of y on X and computes the standard deviation of the regression coefficients
}
\details{
The functions described here are designed as helper functions for the Bag of Little Bootstraps. When used in the BLB, either of these functions can be called on each of the s subsamples, with the matrix X and vector y being a subsample of the full data matrix and response vector.

\code{bootstrap_multidim} runs in serial over the r subsamples.
}
\examples{
X <- matrix(rnorm(10), 5, 2)
y <- rnorm(5)
bootstrap_multidim(X, y, 1000, 30)
}

